Manage folder in src folder was added
