+++ src/Curses/index.jsx
@@ -1,0 +1,19 @@
+import React from 'react'
+import udg from '../assets/images/udg.png'
+
+export default function Curses() {
+  return (
+    <div >
+          <nav class="navbar navbar-light bg-light">
+              <div class="container-fluid">
+                  <a class="navbar-brand" href="#">
+                    <img src={udg} alt="Logo" />;
+                          Bootstrap
+                  </a>
+              </div>
+          </nav>
+    </div>
+  )
+}
+
+
