+++ src/FormRecoverPassword/index.css
@@ -1,0 +1,51 @@
+@import url('https://fonts.googleapis.com/css2?family=Dosis:wght@200&family=Poppins:wght@400;500;600&family=Roboto:wght@300;400;700&display=swap');
+
+body{
+    font-family: 'Roboto', sans-serif;  
+}
+
+.containerLogin {
+    display: flex;
+    min-height: 100vh;
+}
+
+h1{
+    font-size: 1.5rem;
+    font-weight: bold;
+    color: #0C345C;
+    margin: 20px 0px;
+}
+
+h2{
+    font-size: 1rem;
+    margin-bottom: .5rem;
+    color: #0C345C;
+}
+
+form{
+    background-color: white;
+    margin: auto;
+    width: 90px;
+    min-width: 400px;
+    padding: 4.5em 3em;
+    border-radius: 10px;
+    box-shadow: 0 5px 10px -5px rgb( 0 0 0 / 30%);
+    
+}
+
+.buttonForm{
+    margin: 30px;
+    text-align: center;
+}
+
+button {
+    width: 92%;
+    height: 45px;
+    border: none;
+    background-color: #0C345C;
+    color: white;
+    border-radius: 5px;
+    font-size: 18px;
+    cursor: pointer;
+    margin-top: 30px;
+}
