+++ src/FormSignin/index.jsx
@@ -1,9 +1,9 @@
 import {useState} from 'react'
-import FormInput from './FormInput'
+import FormInput from '../FormInput'
 //import "./login2.css"
 import cuceiBlack from '../assets/images/cuceiBlack.png'
 import udg from '../assets/images/udg.png'
-import './styleSignin.css'
+import './index.css'
 
 const FormSignin = () => {
