--- src/components/Xa.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Xa = () => {
-    return (
-        <div>
-            <h1>hellouuuuu</h1>
-        </div>
-    );
-};
-
-export default Xa;
